+++ include/parser/check.hpp
@@ -2,9 +2,13 @@
 
 #include "tokens/tokens.hpp"
 
-namespace check_symbol {
+namespace check::symbol {
 
 	void require(token_list& Token_list, token_symbol::type Type);
 	bool is(token* Token, token_symbol::type Type);
 
+};
+
+namespace check::keyword {
+	bool is(token* Token, token_symbol::type Type);
 };
