+++ include/parser/parse_defination.hpp
@@ -1,4 +1,6 @@
 #pragma once
 
+#include "ast/nodes.hpp"
+
 defination* parse_defination(token_list& Token_list);
 defination_block* parse_defination_block(token_list& Token_list);
