+++ include/parser/parse_if.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+conditional_branches* parse_ifs(token_list& Token_list);
