+++ include/parser/parse_import.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "ast/nodes.hpp"
+
+import* parse_import(token_list& Token_list);
