--- include/parser/parse_keyword.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-statement* parse_statement_keyword(token_list& Token_list);
-
-expr* parse_expression_keyword(token_list& Token_list);
