+++ src/parser/expression/suffix.cpp
@@ -8,13 +8,13 @@ arg_list* parse_callation(token_list& Token_list) {
 	std::vector<expr*> result;
 
 	int begin = Token_list.this_()->BEGIN;
-	while (!check_symbol::is(Token_list.this_(), token_symbol::type::PAREN_RIGHT)) {
+	while (!check::symbol::is(Token_list.this_(), token_symbol::type::PAREN_RIGHT)) {
 		result.push_back(parse_expression(Token_list));
 
-		if (check_symbol::is(Token_list.this_(), token_symbol::type::COMMA)) {
+		if (check::symbol::is(Token_list.this_(), token_symbol::type::COMMA)) {
 			Token_list.next();
 		}
-		else if (check_symbol::is(Token_list.this_(), token_symbol::type::PAREN_RIGHT)) {
+		else if (check::symbol::is(Token_list.this_(), token_symbol::type::PAREN_RIGHT)) {
 			// pass
 		}
 		else {
@@ -30,13 +30,13 @@ expr* parse_suffix(token_list& Token_list, expr* item) {
 
 	while (true) {
 
-		if (check_symbol::is(Token_list.this_(), token_symbol::type::PAREN_LEFT)) {
+		if (check::symbol::is(Token_list.this_(), token_symbol::type::PAREN_LEFT)) {
 			Token_list.next();
 			auto caller = parse_callation(Token_list); // @index+1
 			item = new expr_callation { item, caller, item->BEGIN, caller->END };
 		}
 
-		else if (check_symbol::is(Token_list.this_(), token_symbol::type::DOT)) {
+		else if (check::symbol::is(Token_list.this_(), token_symbol::type::DOT)) {
 			Token_list.next();
 			if (dynamic_cast<token_identifier*> (Token_list.this_()) != NULL) {
 				item = new member_access { item, dynamic_cast<token_identifier*> (Token_list.this_()), item->BEGIN, Token_list.this_()->END };
